Move field type directives into their own directory
