+++ frontend/app/components/wp-edit/field-types/wp-edit-boolean-field.directive.html
