+++ frontend/app/components/wp-edit/field-types/wp-edit-boolean-field.module.ts
@@ -26,8 +26,8 @@
 // See doc/COPYRIGHT.rdoc for more details.
 // ++
 
-import {Field} from "./wp-edit-field.module";
+import {Field} from "../wp-edit-field/wp-edit-field.module";
 
-export class DateField extends Field {
-  public template:string = '/components/wp-edit/wp-edit-field/wp-edit-date-field.directive.html'
+export class BooleanField extends Field {
+  public template:string = '/components/wp-edit/field-types/wp-edit-boolean-field.directive.html'
 }
