+++ frontend/app/components/wp-edit/field-types/wp-edit-date-field.directive.html
