+++ frontend/app/components/wp-edit/field-types/wp-edit-date-field.module.ts
@@ -26,9 +26,8 @@
 // See doc/COPYRIGHT.rdoc for more details.
 // ++
 
-import {Field} from "./wp-edit-field.module";
-import {WorkPackageEditFieldService} from "./wp-edit-field.service";
+import {Field} from "../wp-edit-field/wp-edit-field.module";
 
-export class DurationField extends Field {
-  public template:string = '/components/wp-edit/wp-edit-field/wp-edit-duration-field.directive.html'
+export class DateField extends Field {
+  public template:string = '/components/wp-edit/field-types/wp-edit-date-field.directive.html'
 }
