+++ frontend/app/components/wp-edit/field-types/wp-edit-duration-field.module.ts
@@ -1,3 +1,4 @@
+// -- copyright
 // OpenProject is a project management system.
 // Copyright (C) 2012-2015 the OpenProject Foundation (OPF)
 //
@@ -25,8 +26,8 @@
 // See doc/COPYRIGHT.rdoc for more details.
 // ++
 
-import {Field} from "./wp-edit-field.module";
+import {Field} from "../wp-edit-field/wp-edit-field.module";
 
-export class FloatField extends Field {
-  public template:string = '/components/wp-edit/wp-edit-field/wp-edit-float-field.directive.html'
+export class DurationField extends Field {
+  public template:string = '/components/wp-edit/field-types/wp-edit-duration-field.directive.html'
 }
