+++ frontend/app/components/wp-edit/field-types/wp-edit-float-field.directive.html
