+++ frontend/app/components/wp-edit/field-types/wp-edit-integer-field.directive.html
